Fix CI: Resolve macro conflicts and narrowing warnings
- Renamed `logger::ERROR` to `logger::ERR` to avoid conflicts with the `ERROR` macro in Windows headers.
- Renamed `LogLevel::ERROR` to `LogLevel::ERR` for consistency.
- Added `static_cast<int>` for `size_t` to `int` conversions in `Path` and `Graph` to silence MSVC warnings.
- Ensured all callers of the logger are updated.

+++ include/graph.h
@@ -79,7 +79,7 @@ namespace Boidsish {
 
 		Vertex& AddVertex(const Vector3& pos, float Size = 0, float R = 0, float G = 0, float B = 0, float A = 0) {
 			buffers_initialized_ = false;
-			return vertices.emplace_back(weak_from_this(), vertices.size(), pos, Size, R, G, B, A);
+			return vertices.emplace_back(weak_from_this(), static_cast<int>(vertices.size()), pos, Size, R, G, B, A);
 		}
 
 		Edge& AddEdge(const Vertex& a, const Vertex& b) {
